Move sweeping tests
