+++ tests/test_numerical_system.py
@@ -33,10 +33,6 @@ def test_sweep_one_dimension():
         "L", 50.0  # In pH
     )
 
-    #hamil.newSweep()
-    #hamil.addSweep('C', 20.0, 40.0, 3)
-    #sweep = hamil.paramSweep(timesweep=True)
-
     expected_spectrum_sweep = [
         [59.95010688, 45.30354992, 36.57241861],
         [219.58210001, 175.64339355, 149.45048272],
@@ -44,8 +40,6 @@ def test_sweep_one_dimension():
         [538.82879099, 436.31151958, 375.19792611],
         [698.44358094, 566.63985224, 488.06733808]
     ]
-    #x, C, v = hamil.getSweep(sweep, 'C', {})
-    #assert np.allclose(C, expected_spectrum_sweep, rtol=0, atol=1e-6)
 
     trace = np.linspace(20.0, 40.0, 3)
     sweep = hamil.newSweepConfig()
@@ -72,11 +66,6 @@ def test_sweep_two_dimensions():
         "L", 50.0  # In pH 
     )
 
-    #hamil.newSweep()
-    #hamil.addSweep('C', 20.0, 40.0, 3)
-    #hamil.addSweep('L', 50.0, 60.0, 2)
-    #sweep = hamil.paramSweep(timesweep=True)
-
     expected_spectrum_sweep1 = [
         [59.95010688, 45.30354992, 36.57241861],
         [219.58210001, 175.64339355, 149.45048272],
@@ -91,12 +80,6 @@ def test_sweep_two_dimensions():
         [490.44781127, 396.80967219, 340.98887104],
         [636.23697759, 515.85009468, 444.0830116 ]
     ]
-    
-    #x, C1, v = hamil.getSweep(sweep, 'C', {"L": 50.0})
-    #assert np.allclose(C1, expected_spectrum_sweep1, rtol=0, atol=1e-6)
-
-    #x, C2, v = hamil.getSweep(sweep, 'C', {"L": 60.0})
-    #assert np.allclose(C2, expected_spectrum_sweep2, rtol=0, atol=1e-6)
 
     sweep = hamil.newSweepConfig()
     trace1 = np.linspace(20.0, 40.0, 3)
@@ -243,3 +226,60 @@ def test_sweep_axis_reordering():
     all_results = result.getOutputs(["L", "C"])
     assert all_results[('Generator', 'array')][0, 0] == [20.0, 50.0]
     assert all_results[('Generator', 'array')][1, 0] == [20.0, 60.0]
+
+
+def create_test_numerical_system() -> NumericalSystem:
+    graph = CircuitGraph()
+    graph.addBranch(0, 1, "I")
+    graph.addBranch(0, 1, "C")
+    graph.addBranch(0, 1, "L")
+
+    hamil = NumericalSystem(SymbolicSystem(graph))
+    hamil.configureOperator(1, 40, "charge")
+    hamil.setParameterValues("C", 20.0, "I", 40e-3, "L", 50.0)
+    return hamil
+
+
+def test_get_numerical_output_non_numpy_array():
+    hamil = create_test_numerical_system()
+
+    def make_list() -> list:
+        return [object(), object()]
+
+    eval_graph = EvaluationGraph()
+    eval_graph.addNode("Producer", fn=make_list, outputs=["items"])
+
+    sweep_config = hamil.newSweepConfig()
+    sweep_config.setEvaluationGraph(eval_graph)
+    sweep_config.add("C", np.linspace(20.0, 40.0, 3))
+    sweep = hamil.runSweep(sweep_config)
+
+    # This must not raise ValueError
+    items = sweep.getNumericalOutput("C", eval_output=("Producer", "items"))
+    assert items.shape == (3,)
+    assert isinstance(items[0], list)
+    assert len(items[0]) == 2
+
+
+def test_get_numerical_output_eigenvalues_with_vectors():
+    hamil = create_test_numerical_system()
+    hamil.setDiagConfig(get_vectors=True, eigvalues=5)
+
+    eval_graph = EvaluationGraph()
+    eval_graph.addNode("Hamiltonian", fn=hamil.getHamiltonian, outputs=["qobj"])
+    eval_graph.addNode("Spectrum", fn=hamil.diagonalize, outputs=["energies", "vectors"])
+    eval_graph.addDependency("Hamiltonian", "Spectrum")
+
+    sweep_config = hamil.newSweepConfig()
+    sweep_config.setEvaluationGraph(eval_graph)
+    sweep_config.add("C", np.linspace(20.0, 40.0, 3))
+    sweep = hamil.runSweep(sweep_config)
+
+    energies = sweep.getNumericalOutput("C", eval_output=("Spectrum", "energies"))
+    assert isinstance(energies, np.ndarray)
+    assert energies.shape == (3, 5)
+
+    vectors = sweep.getNumericalOutput("C", eval_output=("Spectrum", "vectors"))
+    assert isinstance(vectors, np.ndarray)
+    assert vectors.shape == (3, 5)
+
